+++ semcore/scroll-area/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.35.0] - 2024-07-31
+
+### Add
+
+- Context to hide scrollbars from screen readers.
+
 ## [5.34.1] - 2024-07-30
 
 ### Changed